Add a dashed orange line showing where moving heads would point if they could move infinitely fast.
